Added more simulation cursors at will.
Added m_signalsGrid context menu entries, Create a new cursor from every cell but Delete only from Cursor 3 and above

New entries are saved into their *.wbk files. Loading of current files should not be affected, as tested,
and on save they get the new property added.

Updates *.wbk file version to "7".

If the tab has more cursors than the loaded file, extra cursors should not be affected, otherwise new cursors shown up.
CursorD left hopefully untouched and kept the legacy behavior close to Cursors 1 & 2.

No more known bugs. It is now working as expected as tested on win11 and Debian 12.

m_cursorsGrid events also work.

Fixes: https://gitlab.com/kicad/code/kicad/-/issues/15211

+++ eeschema/sim/spice_value.cpp
@@ -49,7 +49,7 @@ void SPICE_VALUE_FORMAT::FromString( const wxString& aString )
     {
         long val;
         aString.Left( 1 ).ToLong( &val );
-        Precision = (int) val;
+        Precision = static_cast<int>( val );
         Range = aString.Right( aString.Length() - 1 );
     }
 }
@@ -272,7 +272,7 @@ wxString SPICE_VALUE::ToString( const SPICE_VALUE_FORMAT& aFormat )
     wxString res = wxString::FromCDouble( mantissa, std::max( 0, aFormat.Precision - scale - 1 ) );
 
     // If we have an excessively long number, switch to scientific notation
-    if( ssize_t( res.length() ) > aFormat.Precision + scale + 1 )
+    if( ssize_t( res.length() ) > aFormat.Precision + static_cast<long long>( scale ) + 1 )
         res = wxString::FromCDouble( mantissa );
 
     return res + range;
